refactor(homepage): refactored home page to not combine multiple unrelated views into a single file
